Extract a value container to aiovalues
